+++ package.json
@@ -29,6 +29,7 @@
     "axios": "^0.18.0",
     "highlight.js": "^9.15.6",
     "juice": "^5.2.0",
+    "lodash.debounce": "^4.0.8",
     "markdown-it": "^8.4.2",
     "markdown-it-deflist": "^2.0.3",
     "markdown-it-footnote": "^3.0.1",
@@ -180,9 +181,7 @@
       "jsx",
       "node"
     ],
-    "watchPlugins": [
-
-    ]
+    "watchPlugins": []
   },
   "babel": {
     "presets": [
